pyproject.toml
[project]
name = "discovery-fm"
version = "0.1.0"
description = "Discover music through your Last.fm neighbours - create Spotify playlists from tracks they love that you haven't heard"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pylast>=5.2.0",
    "spotipy>=2.23.0",
    "typer[all]>=0.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]

[project.scripts]
discovery-fm = "discovery_fm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discovery_fm"]
